Module mina_rs_base::account

Module contains the Account data type and associated types

Re-exports

pub use self::zkapp::ZkAppOptionHashableWrapper;

pub use self::zkapp::ZkAppUriOptionHashableWrapper;

pub use permissions::AuthRequired;

pub use permissions::Permissions;

pub use permissions::PermissionsLegacy;

pub use timing::Timing;

pub use token_permissions::TokenPermissions;

pub use token_symbol::TokenSymbol;

pub use zkapp::ZkApp;

pub use zkapp::ZkAppUri;

Modules

permissions

Account based permissions

timing

Account Timing

token_permissions

Account token permissions

token_symbol

Account token symbol

zkapp

zk app

Structs

Account

TODO

AccountLegacy

An account identified by its public key and token ID. Multiple accounts may where the same public key if multiple tokens exist