Trait mina_rs_base::protocol_state::Header

pub trait Header {
    fn get_height(&self) -> Length;
    fn min_window_density(&self) -> Length;
    fn sub_window_densities(&self) -> &Vec<Length>;
}

Implementing types have some notion of height and can return it

Required Methods

fn get_height(&self) -> Length

Get the height for the implementing type

fn min_window_density(&self) -> Length

The minimum window density at the current epoch.

fn sub_window_densities(&self) -> &Vec<Length>

A list of density values of the sub windows.