Module mina_rs_base::staged_ledger_diff

In this context a diff refers to a difference between two states of the blockchain. In this case it is between the current state and the proposed next state.

Structs

CoinBaseBalanceData

CoinBaseFeeTransfer

FeeTransferBalanceData

StagedLedgerDiff

Top level wrapper type for a StagedLedgerDiff

StagedLedgerDiffTuple

StagedLedgerPreDiff

TransactionStatusAuxiliaryData

TransactionStatusBalanceData

UserCommandWithStatus

Enums

CoinBase

InternalCommandBalanceData

TransactionStatus

TransactionStatusFailedType