Module mina_rs_base::user_commands::signed_command

Signed commands are commands that require signing with some accounts private key

Modules

builder

Helpers for building a user command

Structs

SignedCommand

Top level signed command type

SignedCommandPayload

The part of a signed command that needs to be serialized and signed

SignedCommandPayloadCommon

Common fields required by all signed commands

Enums

SignedCommandPayloadBody

Enum of variable fields in a signed command

StakeDelegation

Enum of variable fields for stake delegation