Trait mina_rs_base::verifiable::Verifiable

pub trait Verifiable<CTX> {
    fn verify(&self, ctx: &mut CTX) -> bool;
}

Types that implement Verifiable are able to self-check using intrinsic data (e.g. signatures) and optionally some context that is required for peforming the verification e.g. a singleton signature verifier or param required for snark verification

Required Methods

fn verify(&self, ctx: &mut CTX) -> bool

Accepts self and a context and returns if the implementor is valid

The context must be a mutable reference to satisfy the requirements of using a shared signature checking context which performs self mutation as an optimization