Module mina_rs_base::numbers

Newtypes for different numeric types used throughout Mina

Structs

AccountNonce

Number representing how many txns sent from an account

Amount

This structure represents fixed point numbers typically amounts of Mina currency

BigInt256

Mina 256 bit Bit Integer type

BlockTime

Block time numeric type

BlockTimeSpan

Time span between two block time instants

Char

A single char defined by a single byte (not variable length like a Rust char)

Delta

Represents a difference between two lengths

ExtendedU32

u32 wrapped in 1 version byte This will not be part of the public API once the deserialization refactor is complete

ExtendedU64

This will not be part of the public API once the deserialization refactor is complete

GlobalSlotNumber

Consensus slot index

Hex64

4 bytes wrapped by a version Will not form part of the public API when deserialization refactor is complete

Length

Represents the length of something (e.g. an epoch or window)

MinaIndex

Mina_numbers.Index

SignedAmount

Signed amount, positive -> true, negative -> false

TokenId

Newtype for TokenIds

Enums

ParseAmountError

Error that can be returned when parsing an Amount from string